:recycle: Move dynamoDB subscriptions into handler
