+++ lib/facebook.js
@@ -1,7 +1,6 @@
 const rp = require('request-promise-native');
 const path = require('path');
 const getAttachmentId = require('./facebookAttachments').getAttachmentId;
-const subscriptions = require('../lib/subscriptions');
 
 module.exports.Chat = class Chat {
   constructor(event) {
@@ -134,34 +133,7 @@ module.exports.Chat = class Chat {
             user: this.event.sender.id,
           },
         }).then(() => {
-          console.log(`Labeled ${this.event.sender.id} with ${label}`);
-
-          const item = {
-            morning: (label === 'push-morning'),
-            evening: (label === 'push-evening'),
-          };
-
-          subscriptions.create(this.event.sender.id, item).then(() => {
-            console.log(`Created in dynamoDB ${this.event.sender.id} with ${label}`);
-          }).catch(error => {
-            console.log('Creating user in dynamoDB failed: ' + error);
-
-            let timing = null;
-            if (label === 'push-morning') {
-              timing = 'morning';
-            } else if (label === 'push-evening') {
-              timing = 'evening';
-            } else {
-              console.log('Unknown label:', label);
-              return
-            }
-
-            subscriptions.update(this.event.sender.id, timing, true).then(() => {
-              console.log(`Updated in dynamoDB ${this.event.sender.id} with ${label}`);
-            }).catch(error => {
-              console.log('Updating user in dynamoDB failed: ' + error);
-            });
-          })
+        console.log(`Labeled ${this.event.sender.id} with ${label}`);
       }).catch(error => {
         console.log('Labeling user failed: ' + error);
       })
@@ -184,26 +156,6 @@ module.exports.Chat = class Chat {
           },
         }).then(() => {
         console.log(`Removed label ${label} from ${this.event.sender.id}`);
-
-        let timing = null;
-        if (label === 'push-morning') {
-          timing = 'morning';
-        } else if (label === 'push-evening') {
-          timing = 'evening';
-        } else {
-          console.log('Unknown label:', label);
-          return
-        }
-
-        subscriptions.update(this.event.sender.id, timing, false).then(sub => {
-          if (!sub.morning && !sub.evening) {
-            subscriptions.remove(this.event.sender.id).then(() => {
-              console.log(`Deleted User in dynamoDB with psid ${this.event.sender.id}`);
-            }).catch(error => {
-              console.log(`Deleting user from dynamoDB failed: ${error}`);
-            })
-          }
-        })
       }).catch(error => {
         console.log('Removing label from user failed: ' + error);
       })
